Consolidate Zoom fns
